requirements.txt
openai-whisper==20231117
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.3
ollama==0.1.6
pyttsx3==2.90
torch==2.1.0
torchaudio==2.1.0
